+++ winter_openapi/generator.py
@@ -8,8 +8,12 @@ from typing import List
 from typing import Optional
 from typing import Sequence
 from typing import Set
+from typing import Tuple
+from typing import Type
+from typing import Union
 
 from django.http.response import HttpResponseBase
+from openapi_pydantic import schema_validate
 from openapi_pydantic.v3.v3_0_3 import Components
 from openapi_pydantic.v3.v3_0_3 import Info
 from openapi_pydantic.v3.v3_0_3 import MediaType as MediaTypeModel
@@ -18,12 +22,13 @@ from openapi_pydantic.v3.v3_0_3 import Operation
 from openapi_pydantic.v3.v3_0_3 import Parameter
 from openapi_pydantic.v3.v3_0_3 import PathItem
 from openapi_pydantic.v3.v3_0_3 import Paths
+from openapi_pydantic.v3.v3_0_3 import Reference
 from openapi_pydantic.v3.v3_0_3 import RequestBody
 from openapi_pydantic.v3.v3_0_3 import Response
 from openapi_pydantic.v3.v3_0_3 import Responses
+from openapi_pydantic.v3.v3_0_3 import Schema
 from openapi_pydantic.v3.v3_0_3 import Server
 from openapi_pydantic.v3.v3_0_3 import Tag
-from openapi_pydantic import schema_validate
 
 from winter.core import ComponentMethod
 from winter.web import MediaType
@@ -32,11 +37,45 @@ from winter.web.exceptions import MethodExceptionsManager
 from winter.web.request_body_annotation import RequestBodyAnnotation
 from winter.web.routing import Route
 from winter.web.routing import RouteAnnotation
-from winter_openapi.inspection.inspection import InspectorNotFound
 from winter_openapi.inspection.inspection import inspect_type
+from .inspection.inspection import InspectorNotFound
 from .inspectors import get_route_parameters_inspectors
 from .type_info_converter import convert_type_info_to_openapi_schema
 
+
+class SchemaRegistry:
+    def __init__(self):
+        self._schemas: Dict[Tuple[Type, bool], Schema] = {}
+        self._types_by_titles: Dict[str, Type] = {}
+
+    def get_schema_or_reference(self, type_: Type, output: bool) -> Union[Schema, Reference]:
+        type_info = inspect_type(type_)
+        schema = self._schemas.get((type_, output))
+        if not schema:
+            schema = convert_type_info_to_openapi_schema(type_info, output=output, schema_registry=self)
+            if not schema.title:
+                return schema
+            self._schemas[type_, output] = schema
+
+        if schema.title not in self._types_by_titles:
+            self._types_by_titles[schema.title] = type_
+        elif self._types_by_titles[schema.title] != type_:
+            raise ValueError(f'Title {schema.title} for type {type_} is already used for another type: {self._types_by_titles[schema.title]}')
+
+        reference = Reference(ref='#/components/schemas/' + schema.title)
+
+        if type_info.nullable and schema.type == 'object':
+            # https://stackoverflow.com/questions/40920441/how-to-specify-a-property-can-be-null-or-a-reference-with-swagger
+            # Better solution, but not implemented yet https://github.com/OpenAPITools/openapi-generator/issues/9083
+            return Schema(nullable=True, allOf=[reference])
+
+        return reference
+
+    def get_schemas(self) -> Dict[str, Schema]:
+        return {
+            schema.title: schema
+            for schema in self._schemas.values()
+        }
 
 def generate_openapi(
     title: str,
@@ -49,7 +88,7 @@ def generate_openapi(
 ) -> Dict[str, Any]:
     routes = list(routes)
     routes.sort(key=lambda r: r.url_path)
-    components = Components(responses=[], schemas=[], parameters=[])
+    schema_registry = SchemaRegistry()
     tags_ = [Tag(**tag) for tag in tags or []]
     tag_names = [tag_.name for tag_ in tags_]
     paths: Paths = {}
@@ -68,11 +107,21 @@ def generate_openapi(
             if url_path_tag:
                 path_tag_names.append(url_path_tag)
 
-        path_item = _get_openapi_path(routes=group_routes, operation_ids=operation_ids, tag_names=path_tag_names)
+        path_item = _get_openapi_path(
+            routes=group_routes,
+            operation_ids=operation_ids,
+            tag_names=path_tag_names,
+            schema_registry=schema_registry,
+        )
         paths[url_path_without_prefix] = path_item
 
     info = Info(title=title, version=version, description=description)
     servers_ = [Server(url=path_prefix)]
+    components = Components(
+        schemas=schema_registry.get_schemas(),
+        responses={},
+        parameters={},
+    )
     openapi = OpenAPI(info=info, servers=servers_, paths=paths, components=components, tags=tags_)
     openapi_dict = openapi.dict(by_alias=True, exclude_none=True)
     if validate:
@@ -80,26 +129,46 @@ def generate_openapi(
     return openapi_dict
 
 
-def _get_openapi_path(*, routes: Iterable[Route], operation_ids: Set[str], tag_names: Iterable[str]) -> PathItem:
+def _get_openapi_path(
+    *,
+    routes: Iterable[Route],
+    operation_ids: Set[str],
+    tag_names: Iterable[str],
+    schema_registry: SchemaRegistry,
+) -> PathItem:
     path = {}
-    for route_item in routes:
-        operation_id = route_item.method.full_name
+    for route in routes:
+        operation_id = route.method.full_name
         if operation_id in operation_ids:
             warnings.warn(f"Duplicate Operation ID {operation_id}")
         operation_ids.add(operation_id)
 
-        operation = _get_openapi_operation(route=route_item, operation_id=operation_id, tag_names=tag_names)
-        path[route_item.http_method.lower()] = operation
+        try:
+            operation = _get_openapi_operation(
+                route=route,
+                operation_id=operation_id,
+                tag_names=tag_names,
+                schema_registry=schema_registry,
+            )
+        except InspectorNotFound as e:
+            raise CanNotInspectType(route.method, str(e))
+        path[route.http_method.lower()] = operation
 
     return PathItem.parse_obj(path)
 
 
-def _get_openapi_operation(*, route: Route, operation_id: str, tag_names: Iterable[str]) -> Operation:
+def _get_openapi_operation(
+    *,
+    route: Route,
+    operation_id: str,
+    tag_names: Iterable[str],
+    schema_registry: SchemaRegistry,
+) -> Operation:
     summary = route.method.docstring.short_description
     description = route.method.docstring.long_description
-    operation_parameters = get_route_parameters(route)
-    operation_request_body = get_request_body_parameters(route)
-    operation_responses = get_responses_schemas(route)
+    operation_parameters = get_route_parameters(route, schema_registry)
+    operation_request_body = get_request_body_parameters(route, schema_registry)
+    operation_responses = get_responses_schemas(route, schema_registry)
     return Operation(
         tags=tag_names,
         summary=summary,
@@ -111,15 +180,13 @@ def _get_openapi_operation(*, route: Route, operation_id: str, tag_names: Iterab
     )
 
 
-class CanNotInspectReturnType(Exception):
+class CanNotInspectType(Exception):
 
     def __init__(
         self,
         method: ComponentMethod,
-        return_type: Any,
         message: str,
     ):
-        self._return_type = return_type
         self._message = message
         self._method = method
 
@@ -129,7 +196,7 @@ class CanNotInspectReturnType(Exception):
     def __str__(self):
         component_cls = self._method.component.component_cls
         method_path = f'{component_cls.__module__}.{self._method.full_name}'
-        return f'{method_path}: -> {self._return_type}: {self._message}'
+        return f'{method_path}: {self._message}'
 
 
 def get_url_path_without_prefix(url_path: str, path_prefix: str) -> str:
@@ -163,14 +230,14 @@ def get_url_path_tag(url_path: str, path_prefix: str) -> Optional[str]:
     return url_path_tag
 
 
-def get_route_parameters(route: Route) -> List[Parameter]:
+def get_route_parameters(route: Route, schema_registry: SchemaRegistry) -> List[Parameter]:
     parameters = []
     for inspector in get_route_parameters_inspectors():
-        parameters += inspector.inspect_parameters(route)
+        parameters += inspector.inspect_parameters(route, schema_registry)
     return parameters
 
 
-def get_request_body_parameters(route: Route) -> Optional[RequestBody]:
+def get_request_body_parameters(route: Route, schema_registry: SchemaRegistry) -> Optional[RequestBody]:
     method = route.method
     request_body_annotation = method.annotations.get_one_or_none(RequestBodyAnnotation)
     if request_body_annotation is None:
@@ -178,31 +245,30 @@ def get_request_body_parameters(route: Route) -> Optional[RequestBody]:
 
     description = method.docstring.short_description
     argument = method.get_argument(request_body_annotation.argument_name)
-    type_info = inspect_type(argument.type_)
-    media_type_schema = convert_type_info_to_openapi_schema(type_info, output=False)
+    reference = schema_registry.get_schema_or_reference(argument.type_, output=False)
 
     route_annotation = method.annotations.get_one_or_none(RouteAnnotation)
     consumes = route_annotation.consumes or [MediaType.APPLICATION_JSON]
     content = {
-        str(consume): MediaTypeModel(media_type_schema=media_type_schema)
+        str(consume): MediaTypeModel(media_type_schema=reference)
         for consume in consumes
     }
     return RequestBody(description=description, content=content)
 
 
-def get_responses_schemas(route: Route) -> Responses:
+def get_responses_schemas(route: Route, schema_registry: SchemaRegistry) -> Responses:
     responses: Responses = {}
     http_method = route.http_method
     response_status = str(get_response_status(http_method, route.method))
 
-    responses[response_status] = _build_response_schema(route.method)
+    responses[response_status] = _build_response_schema(route.method, schema_registry)
     method_exceptions_manager = MethodExceptionsManager(route.method)
 
     for exception_cls in method_exceptions_manager.declared_exception_classes:
         handler = method_exceptions_manager.get_handler(exception_cls)
         handle_method = ComponentMethod.get_or_create(handler.__class__.handle)
         response_status = str(get_response_status(http_method, handle_method))
-        responses[response_status] = _build_response_exception_handler_schema(handle_method)
+        responses[response_status] = _build_response_exception_handler_schema(handle_method, schema_registry)
     return responses
 
 
@@ -258,39 +324,29 @@ def determine_path_prefix(url_paths: List[str]) -> str:
     return '/' + '/'.join(common)
 
 
-def _build_response_schema(method: ComponentMethod) -> Response:
+def _build_response_schema(method: ComponentMethod, schema_registry: SchemaRegistry) -> Response:
     return_value_type = method.return_value_type
     if _is_abstract_or_none_return_type(return_value_type):
         return Response(description='')
 
-    type_info = _get_inspect_type(method, return_value_type)
-    media_type_schema = convert_type_info_to_openapi_schema(type_info, output=True)
+    reference = schema_registry.get_schema_or_reference(return_value_type, output=True)
     route_annotation = method.annotations.get_one_or_none(RouteAnnotation)
     produces = route_annotation.produces or [MediaType.APPLICATION_JSON]
     content = {
-        str(produce): MediaTypeModel(media_type_schema=media_type_schema)
+        str(produce): MediaTypeModel(media_type_schema=reference)
         for produce in produces
     }
     return Response(description='', content=content)
 
 
-def _build_response_exception_handler_schema(method: ComponentMethod) -> Response:
+def _build_response_exception_handler_schema(method: ComponentMethod, schema_registry: SchemaRegistry) -> Response:
     return_value_type = method.return_value_type
     if _is_abstract_or_none_return_type(return_value_type):
         return Response(description='')
 
-    type_info = _get_inspect_type(method, return_value_type)
-    media_type_schema = convert_type_info_to_openapi_schema(type_info, output=True)
-    content = {str(MediaType.APPLICATION_JSON): MediaTypeModel(media_type_schema=media_type_schema)}
+    reference = schema_registry.get_schema_or_reference(return_value_type, output=True)
+    content = {str(MediaType.APPLICATION_JSON): MediaTypeModel(media_type_schema=reference)}
     return Response(description='', content=content)
-
-
-def _get_inspect_type(method, return_value_type):
-    try:
-        type_info = inspect_type(return_value_type)
-    except InspectorNotFound as e:
-        raise CanNotInspectReturnType(method, return_value_type, str(e))
-    return type_info
 
 
 def _is_abstract_or_none_return_type(return_value_type):