+++ winter_openapi/inspectors/path_parameters_inspector.py
@@ -1,40 +1,36 @@
 from typing import List
+from typing import TYPE_CHECKING
 
 from openapi_pydantic.v3.v3_0_3 import Parameter
 
 from winter.core import ComponentMethodArgument
 from winter.web.routing import Route
-from winter_openapi.inspection.inspection import InspectorNotFound
-from winter_openapi.inspection.inspection import inspect_type
-from winter_openapi.type_info_converter import convert_type_info_to_openapi_schema
 from .route_parameters_inspector import RouteParametersInspector
-from ..inspection import DataTypes
-from ..inspection import TypeInfo
+
+if TYPE_CHECKING:
+    from winter_openapi.generator import SchemaRegistry
 
 
 class PathParametersInspector(RouteParametersInspector):
 
-    def inspect_parameters(self, route: 'Route') -> List[Parameter]:
+    def inspect_parameters(self, route: 'Route', schema_registry: 'SchemaRegistry') -> List[Parameter]:
         parameters = []
 
         for argument in self._path_arguments(route):
-            openapi_parameter = self._convert_argument_to_openapi_parameter(argument)
+            openapi_parameter = self._convert_argument_to_openapi_parameter(argument, schema_registry)
             parameters.append(openapi_parameter)
 
         return parameters
 
-    def _convert_argument_to_openapi_parameter(self, argument: ComponentMethodArgument) -> Parameter:
-        try:
-            type_info = inspect_type(argument.type_)
-            description = argument.description
-        except InspectorNotFound:
-            type_info = TypeInfo(DataTypes.STRING)
-            description = 'winter_openapi has failed to inspect the parameter'
-
-        schema = convert_type_info_to_openapi_schema(type_info, output=False)
+    def _convert_argument_to_openapi_parameter(
+        self,
+        argument: ComponentMethodArgument,
+        schema_registry: 'SchemaRegistry',
+    ) -> Parameter:
+        schema = schema_registry.get_schema_or_reference(argument.type_, output=False)
         return Parameter(
             name=argument.name,
-            description=description,
+            description=argument.description,
             required=argument.required,
             param_in='path',
             default=argument.get_default(None),