+++ winter_openapi/type_info_converter.py
@@ -1,10 +1,14 @@
+from typing import TYPE_CHECKING
+
 from openapi_pydantic.v3.v3_0_3 import Schema
 
 from winter_openapi.inspection import DataTypes
 from winter_openapi.inspection import TypeInfo
 
+if TYPE_CHECKING:
+    from winter_openapi.generator import SchemaRegistry
 
-def convert_type_info_to_openapi_schema(value: TypeInfo, *, output: bool) -> Schema:
+def convert_type_info_to_openapi_schema(value: TypeInfo, *, output: bool, schema_registry: 'SchemaRegistry') -> Schema:
     if value.type_ == DataTypes.ANY:
         return Schema(
             description='Can be any value - string, number, boolean, array or object.',
@@ -12,8 +16,12 @@ def convert_type_info_to_openapi_schema(value: TypeInfo, *, output: bool) -> Sch
         )
 
     data = {
-        'type': value.type_.value
+        'type': value.type_,
     }
+
+    if value.type_ != DataTypes.OBJECT and value.nullable:
+        data['nullable'] = True
+
     if value.title:
         data['title'] = value.title if output else f'{value.title}Input'
 
@@ -21,13 +29,10 @@ def convert_type_info_to_openapi_schema(value: TypeInfo, *, output: bool) -> Sch
         data['description'] = value.description
 
     if value.format_ is not None:
-        data['schema_format'] = value.format_.value
+        data['schema_format'] = value.format_
 
     if value.child is not None:
-        data['items'] = convert_type_info_to_openapi_schema(value.child, output=output)
-
-    if value.nullable:
-        data['nullable'] = True
+        data['items'] = schema_registry.get_schema_or_reference(value.child.hint_class, output=output)
 
     if value.enum is not None:
         data['enum'] = value.enum
@@ -35,7 +40,7 @@ def convert_type_info_to_openapi_schema(value: TypeInfo, *, output: bool) -> Sch
     if value.properties:
         sorted_keys = sorted(value.properties.keys())
         data['properties'] = {
-            key: convert_type_info_to_openapi_schema(value.properties[key], output=output)
+            key: schema_registry.get_schema_or_reference(value.properties[key].hint_class, output=output)
             for key in sorted_keys
         }