+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [25.0.0] - 2024-07-26
+- Schemas extracted to OpenAPI components, are always referenced
+- Better nullable handling in OpenAPI schemas
+
 ## [24.0.1] - 2024-07-23
 - Added support of Python 3.12
 - Replaced openapi-schema-pydantic with openapi-pydantic